Changed the way to get the samples of the ADC ESP32.
Implemented the root-mean-square current calculation:

Each sample of the ADC pass through in a digital low filter to
extract the 2.5V or 1.65V DC offset, then substract this to centered on 0 counts.
After this, the root-mean-square method current is applied.
This cycle repeat until reach the number of samples.
After this, return the root-mean-square current value.

Added a calibration value and a adjust loop in the begining of the process
to stabilize the sensor data sampling values.

Implemented the kWh calculation.

Implemented the cost per kWh based on 2020 CPFL tarif.